回款金额条数

+++ erui-parent/erui-order/src/main/java/com/erui/order/model/ProjectStatistics.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import org.apache.commons.lang3.StringUtils;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 
 /**
@@ -97,7 +98,7 @@ public class ProjectStatistics {
     // 回款金额（美元）
     private BigDecimal money;   //回款金额
     //回款记录条数
-    private Integer accountCount; //回款记录
+    private BigInteger accountCount; //回款记录
     //初步利润率
     private BigDecimal profitPercent;
     //授信情况
@@ -468,11 +469,11 @@ public class ProjectStatistics {
         this.exeChgDate = exeChgDate;
     }
 
-    public Integer getAccountCount() {
+    public BigInteger getAccountCount() {
         return accountCount;
     }
 
-    public void setAccountCount(Integer accountCount) {
+    public void setAccountCount(BigInteger accountCount) {
         this.accountCount = accountCount;
     }
 }
